Cleaned up commandline arguments.

+++ scrape_files.py
@@ -3,7 +3,7 @@ import argparse
 import fnmatch
 import os
 import sys
-import ipdb
+#import ipdb
 
 from urllib2 import urlopen
 from lxml.html import parse
@@ -34,33 +34,74 @@ def get_csv_urls(url, prefix=None, pattern='*party_[0-9]*.csv'):
     return csv_hrefs
 
 if __name__ == '__main__':
-    #url = "http://www.electionresults.govt.nz/electionresults_XXXX/e9/html/e9_part8.html"             
-
          
     # Parse the commandline arguments
+    general_description = '''
+    The script is used to download New Zealand election results 
+    form the New Zealand Electrol Commission http://www.elections.org.nz/.
+    A URL or a year of the webpage or a year can be entered which will use 
+    a predefined URL for the results
+    '''
     parser = argparse.ArgumentParser(
-        description="For downloading files listed in a webpage")
+        description=general_description)
 
+    pattern_help="""Pattern of the files you want to download in quotes
+    uses globbing to select the files to download e.g. '*.csv' will download 
+    all the csv files from the webpage.
+    """
     parser.add_argument('pattern', type=str,  
-                        metavar='pattern of files to download must be in quotes', 
+                        help=pattern_help,
                         action='store')
-    parser.add_argument('url', action='store')
-    parser.add_argument('-d', '--directory', dest='directory',
-                        metavar='output directory', default=os.getcwd())
+    dir_help = '''Output directory for the downloaded files. 
+    Will create the directory if it already does not exist.
+    '''
+    parser.add_argument('-d', '--directory',
+                        help='output directory for the files', 
+                        default=os.getcwd())
+
+    group = parser.add_mutually_exclusive_group(required=True)
+    url_help = 'URL of the webpage where the links to the files are'
+    group.add_argument('-u', '--url', dest='url',
+                       help=url_help)
+    year_help = '''Year of general election held will use a predefined URL
+    for the votes at each polling booth.'''
+    group.add_argument('-y', '--year', dest='year',
+                       choices=['2014', '2011', '2008', '2005', '2002'],
+                       help=year_help)
 
     args = parser.parse_args()
-    print(args.directory)
-    print(args.pattern)
-    print(args.url)
-    #sys.exit()
-    prefix = args.url.rstrip(args.url.split("/")[-1])
+    #print(args.directory)
+    #print(args.pattern)
+    #print(args.url)
 
-    for link in get_csv_urls(args.url, prefix):
+    if args.directory:
+        if not os.path.exists(args.directory):
+            os.makedirs(args.directory)
+
+    if args.url:
+        prefix = args.url.rstrip(args.url.split("/")[-1])
+        url = args.url
+
+    if args.year:
+        if args.year == '2014':
+            url = ('http://www.electionresults.govt.nz/'
+                   'electionresults_2014/e9/html/e9_part8_party_index.html')
+        else:
+           url = ('http://www.electionresults.govt.nz/'
+                  'electionresults_XXXX/e9/html/e9_part8.html')
+           url = url.replace('XXXX', args.year)
+           prefix = url.rstrip(url.split("/")[-1])
+
+
+    for link in get_csv_urls(url, prefix):
         file_name = link.split("/")[-1]
-        csv_file = urlopen(link)
-        # Open our file
-        with open(file_name, 'w') as f:
-            f.write(csv_file.read())
-        print(file_name)
+
+        # Check if the file exists and if not open the url and write the file 
+        file_name = os.path.join(args.directory, file_name)
+        if not os.path.isfile(file_name):
+            csv_file = urlopen(link)
+            with open(file_name, 'w') as f:
+                f.write(csv_file.read())
+                print(file_name)
